started refactoring process, routes -> app.js
